refactor(db): extract relations into schema.relations.ts
Move all 21 Drizzle relation definitions from schema.ts into a new
schema.relations.ts file, re-exported via schema.ts. This enables
future PostgreSQL schema files to share the same dialect-agnostic
relation definitions.
